Reworked Readme Structure
reworked the readme structure to no longer have a specific feature list (as that is not yet extensively planned) and to instead have a general description of the addin and its features. Additionally two new sections were added highlighting the Target audience and the potential use cases of the addin.